# docs/swagger/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "phpList4 REST API",
        "version": "4.0.0"
    },
    "paths": {
...
        "/api/v2/sessions": {
            "post": {
                "tags": [
                    "sessions"
                ],
                "summary": "Log in or create new session.",
                "description": "Given valid login data, this will generate a login token that will be valid for 1 hour",
                "operationId": "PhpList\\RestBundle\\Controller\\SessionController::postAction",
                "requestBody": {
                    "description": "Pass session credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login_name",
                                    "password"
                                ],
                                "properties": {
                                    "login_name": {
                                        "description": "Creates a new session (if the provided credentials are valid).",
                                        "type": "string",
                                        "format": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "description": "Creates a new session (if the provided credentials are valid).",
                                        "type": "string",
                                        "format": "password",
                                        "example": "eetIc/Gropvoc1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Creates a new session (if the provided credentials are valid).",
                                            "type": "integer",
                                            "example": "1234"
                                        },
                                        "key": {
                                            "description": "Creates a new session (if the provided credentials are valid).",
                                            "type": "string",
                                            "example": "2cfe100561473c6cdd99c9e2f26fa974"
                                        },
                                        "expiry": {
                                            "description": "Creates a new session (if the provided credentials are valid).",
                                            "type": "string",
                                            "example": "2017-07-20T18:22:48+00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Creates a new session (if the provided credentials are valid).",
                                            "type": "string",
                                            "example": "Empty json, invalid data and or incomplete data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Creates a new session (if the provided credentials are valid).",
                                            "type": "string",
                                            "example": "Not authorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/subscriber": {
            "post": {
                "tags": [
                    "subscribers"
                ],
                "summary": "Create a subscriber list",
                "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\n     *     address yet).",
                "operationId": "PhpList\\RestBundle\\Controller\\SubscriberController::postAction",
                "parameters": [
                    {
                        "name": "session",
                        "in": "header",
                        "description": "Session ID obtained from authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass session credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                        "type": "string",
                                        "format": "string",
                                        "example": "admin"
                                    },
                                    "confirmed": {
                                        "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                        "type": "string",
                                        "format": "boolean",
                                        "example": "eetIc/Gropvoc1"
                                    },
                                    "blacklisted": {
                                        "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                        "type": "string",
                                        "format": "boolean",
                                        "example": "eetIc/Gropvoc1"
                                    },
                                    "html_entail": {
                                        "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                        "type": "string",
                                        "format": "boolean",
                                        "example": "eetIc/Gropvoc1"
                                    },
                                    "disabled": {
                                        "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                        "type": "string",
                                        "format": "boolean",
                                        "example": "eetIc/Gropvoc1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "creation_date": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "integer",
                                            "example": "2017-12-16T18:44:27+00:00"
                                        },
                                        "email": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "string",
                                            "example": "subscriber@example.com"
                                        },
                                        "confirmed": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "blacklisted": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "bounced": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "unique_id": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "string",
                                            "example": "69f4e92cf50eafca9627f35704f030f4"
                                        },
                                        "html_entail": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "disabled": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "id": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "No valid session key was provided as basic auth password."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "string",
                                            "example": "This resource already exists."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Creates a new subscriber (if the provided data is valid and there is no subscriber with the given email\naddress yet).",
                                            "type": "string",
                                            "example": "Some fields invalid: email, confirmed, html_email"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}